Add changes for code quality
