--- src/main/java/alcazar/Exceptions/AlcazarException.java
@@ -1,9 +1,0 @@
-package alcazar.Exceptions;
-
-import alcazar.Alcazar;
-
-public class AlcazarException extends Exception {
-    public AlcazarException(String msg) {
-        super(msg);
-    }
-}
