--- src/main/java/alcazar/Exceptions/InvalidArgumentException.java
@@ -1,6 +1,0 @@
-package alcazar.Exceptions;
-public class InvalidArgumentException extends AlcazarException {
-    public InvalidArgumentException(String msg) {
-        super(msg);
-    }
-}
