--- src/main/java/alcazar/Exceptions/InvalidSerialException.java
@@ -1,6 +1,0 @@
-package alcazar.Exceptions;
-public class InvalidSerialException extends AlcazarException{
-    public InvalidSerialException(String msg) {
-        super(msg);
-    }
-}
