--- src/main/java/alcazar/Exceptions/InvalidTaskException.java
@@ -1,6 +1,0 @@
-package alcazar.Exceptions;
-public class InvalidTaskException extends AlcazarException {
-    public InvalidTaskException(String msg) {
-        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
-    }
-}
