+++ src/main/java/alcazar/tasks/Task.java
@@ -1,15 +1,25 @@
-package alcazar.Tasks;
+package alcazar.tasks;
+
+/**
+ * Encapsulates a Task
+ */
 public class Task {
+    /** Boolean showing whether this task has been marked */
     private boolean isMarked;
+    /** The input prompt describing this task */
     private String taskPrompt;
 
+    /**
+     * Constructs a new Task
+     * @param taskPrompt Input description of this task
+     */
     Task(String taskPrompt) {
         this.taskPrompt = taskPrompt;
         this.isMarked = false;
     }
 
     /**
-     * Used to mark this task done.
+     * Marks this task done.
      */
     public void markTask() {
         this.isMarked = true;
@@ -22,7 +32,7 @@ public class Task {
         this.isMarked = false;
     }
     /**
-     * Used to evaluate String form of a Task
+     * Evaluates String form of a Task
      * @return String form of a Task
      */
     public String toString() {
