+++ src/main/java/alcazar/tasks/ToDo.java
@@ -1,10 +1,13 @@
-package alcazar.Tasks;
+package alcazar.tasks;
 
-import alcazar.Tasks.Task;
-
+/**
+ * Encapsulates a ToDo Task
+ */
 public class ToDo extends Task {
-
-
+    /**
+     * Constructs a new ToDo task
+     * @param description The input description of this ToDo
+     */
     public ToDo(String description) {
         super(description);
     }
@@ -13,8 +16,6 @@ public class ToDo extends Task {
      * Evaluates String form of a ToDo task
      * @return String form of a ToDo Task
      */
-
-
     @Override
     public String toString() {
         return "[T]" + super.toString();
