+++ src/test/java/alcazar/AddEventCommandTest.java
@@ -1,0 +1,2 @@
+package alcazar;public class AddEventCommandTest {
+}
